django-app/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
pyephem==4.1.4
geopy==2.4.0
